Move A1 exceptions to utils.exceptions
