+++ src/multi_vector_simulator/A1_csv_to_json.py
@@ -75,42 +75,19 @@ from multi_vector_simulator.utils.constants_json_strings import (
     SOC_MIN,
     STORAGE_CAPACITY,
     FILENAME,
-)
-from multi_vector_simulator.utils.constants_json_strings import (
     UNIT,
     VALUE,
     ENERGY_STORAGE,
 )
 
 
-class MissingParameterError(ValueError):
-    """Exception raised for missing parameters of a csv input file."""
-
-    pass
-
-
-class MissingParameterWarning(UserWarning):
-    """Exception raised for missing new parameters of a csv input file, which will be set to default."""
-
-    pass
-
-
-class WrongParameterWarning(UserWarning):
-    """Exception raised for errors in the parameters of a csv input file."""
-
-    pass
-
-
-class CsvParsingError(ValueError):
-    """Exception raised for errors in the parameters of a csv input file."""
-
-    pass
-
-
-class WrongStorageColumn(ValueError):
-    """Exception raised for wrong column name in "storage_xx" input file."""
-
-    pass
+from multi_vector_simulator.utils.exceptions import (
+    MissingParameterError,
+    MissingParameterWarning,
+    WrongParameterWarning,
+    CsvParsingError,
+    WrongStorageColumn,
+)
 
 
 def create_input_json(
